Move constants into interfaces instead of init
