+++ src/zeit/importer/__init__.py
@@ -1,3 +1,1 @@
-DOC_NS = 'http://namespaces.zeit.de/CMS/document'
-PRINT_NS = 'http://namespaces.zeit.de/CMS/print'
-WORKFLOW_NS = 'http://namespaces.zeit.de/CMS/workflow'
+# python package
