+++ src/zeit/importer/interfaces.py
@@ -1,5 +1,10 @@
 import zope.interface
 
 
+DOC_NS = 'http://namespaces.zeit.de/CMS/document'
+PRINT_NS = 'http://namespaces.zeit.de/CMS/print'
+WORKFLOW_NS = 'http://namespaces.zeit.de/CMS/workflow'
+
+
 class ISettings(zope.interface.Interface):
     pass
